Move appstateservice to data
