+++ lib/ui/route/loading.dart
@@ -1,14 +1,11 @@
-import 'dart:async';
-
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/scheduler.dart';
+import 'package:genshin_mod_manager/domain/repo/app_state_service.dart';
 import 'package:genshin_mod_manager/ui/constant.dart';
-import 'package:genshin_mod_manager/ui/service/app_state_service.dart';
 import 'package:genshin_mod_manager/ui/widget/appbar.dart';
 import 'package:go_router/go_router.dart';
 import 'package:logger/logger.dart';
 import 'package:provider/provider.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:window_manager/window_manager.dart';
 
 class LoadingRoute extends StatelessWidget {
@@ -19,11 +16,8 @@ class LoadingRoute extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final initFuture =
-        context.select<AppStateService, Future<SharedPreferences>>(
-            (value) => value.initFuture);
     return FutureBuilder(
-      future: initFuture,
+      future: context.select((AppStateService value) => value.successfulLoad),
       builder: (context, snapshot) {
         if (snapshot.connectionState != ConnectionState.done) {
           return _buildLoading();
