+++ lib/ui/router.dart
@@ -1,6 +1,8 @@
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:genshin_mod_manager/data/extension/pathops.dart';
+import 'package:genshin_mod_manager/data/repo/app_state_service.dart';
 import 'package:genshin_mod_manager/domain/entity/mod_category.dart';
+import 'package:genshin_mod_manager/domain/repo/app_state_service.dart';
 import 'package:genshin_mod_manager/ui/constant.dart';
 import 'package:genshin_mod_manager/ui/route/category/category.dart';
 import 'package:genshin_mod_manager/ui/route/home_shell/home_shell.dart';
@@ -9,7 +11,6 @@ import 'package:genshin_mod_manager/ui/route/loading.dart';
 import 'package:genshin_mod_manager/ui/route/nahida_store/nahida_store.dart';
 import 'package:genshin_mod_manager/ui/route/setting/setting.dart';
 import 'package:genshin_mod_manager/ui/route/welcome.dart';
-import 'package:genshin_mod_manager/ui/service/app_state_service.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 
@@ -72,7 +73,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider(
-      create: (context) => AppStateService(),
+      create: (context) => getAppStateService(),
       child: FluentApp.router(
         title: 'Genshin Mod Manager',
         routerDelegate: _router.routerDelegate,
