+++ lib/ui/widget/category_drop_target.dart
@@ -4,7 +4,7 @@ import 'package:desktop_drop/desktop_drop.dart';
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:genshin_mod_manager/data/extension/copy_directory.dart';
 import 'package:genshin_mod_manager/data/extension/pathops.dart';
-import 'package:genshin_mod_manager/ui/service/app_state_service.dart';
+import 'package:genshin_mod_manager/domain/repo/app_state_service.dart';
 import 'package:logger/logger.dart';
 import 'package:provider/provider.dart';
